requirements.txt
fastapi[all]~=0.115.11
fastapi-users[sqlalchemy]~=14.0.1
fastapi-cache2[redis]~=0.2.2
sqlalchemy~=2.0.39
alembic~=1.15.1
asyncpg~=0.30.0
psycopg2-binary~=2.9.10
redis~=4.6.0
aiosqlite~=0.21.0
python-dotenv~=1.0.1
pydantic~=2.10.6
uvicorn~=0.34.0
gunicorn~=23.0.0
httpx~=0.28.1
pytest~=8.3.5
pytest-asyncio~=0.26.0
pytest-mock~=3.14.0
pytest-cov~=6.0.0
coverage~=7.8.0
locust~=2.33.2
